ros2_ws/fleet_ui/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pyyaml>=6.0
